+++ zcash_proofs/src/sapling/mod.rs
@@ -1,3 +1,0 @@
-//! Helpers for creating Sapling proofs.
-
-mod prover;